Simplify zune build and combine main
