+++ src/commands/test.zig
@@ -2,7 +2,8 @@ const std = @import("std");
 const luau = @import("luau");
 const builtin = @import("builtin");
 
-const Zune = @import("../zune.zig");
+const Zune = @import("zune");
+
 const command = @import("lib.zig");
 
 const Engine = @import("../core/runtime/engine.zig");
