+++ src/core/utils/testrunner.zig
@@ -3,7 +3,8 @@ const xev = @import("xev");
 const luau = @import("luau");
 const builtin = @import("builtin");
 
-const Zune = @import("../../zune.zig");
+const Zune = @import("zune");
+
 const Engine = @import("../runtime/engine.zig");
 const Scheduler = @import("../runtime/scheduler.zig");
