+++ test/standard/ffi/init.test.luau
@@ -7,6 +7,10 @@ local test = testing.test;
 local expect = testing.expect;
 local describe = testing.describe;
 
+local function warn(message: string)
+    print("\27[33m" .. message .. "\27[0m");
+end
+
 if (not ffi) then
     warn("Skipping 'FFI' tests (Not Supported)");
     return;
@@ -60,7 +64,11 @@ else
     error("Unsupported OS: " .. process.os);
 end
 
+-- Workaround for Windows tests, test sample library
 local static_suffix = ffi.suffix;
+if (process.os == "windows") then
+    -- static_suffix = "lib";
+end
 
 local function testBufferOfValue(size: number, writer: (b: buffer, o: number, v: number) -> (), value: number): buffer
     local buf = buffer.create(size);
@@ -1116,6 +1124,7 @@ describe("FFI", function()
                 local ptr = nil;
                 test("Symbol<*i32 new_i32()>", function()
                     ptr = lib.new_i32();
+                    print("new_i32");
                     ptr:setSize(ffi.types.i32:size());
                     expect(ptr).toBe(expect.type("userdata"));
